Add multiflag support for new flag implementation
In the parent of this commit, I've added a new flag implementation to
support config using env vars or a config file. This commits add support
for multiflag value setting, when using a csv like structure for those
values.

This was needed as for those only the first value was being read.