Create /dev/random and /dev/urandom when daemonizing and jailing
Go on Linux normally uses the getrandom() syscall to get entropy. However,
this is onl available for Linux version >= 3.17

Making /dev/random and /dev/urandom available in the chrooted jail allows the
daemon to operate correctly on these systems. It's important to handle this
automatically, as the previous workaround of manually creating these character
devices is made more difficult by the use of a random directory in $TMPDIR.