Disable automatic HTTP -> HTTPS redirection if no HTTPS server defined
Right now the logic for determing when HTTPS redirection should happen
is split in both the GitLab Rails and Pages processes. This means that
the Pages configuration in Omnibus has to be specified correctly
for Sidekiq for the HTTPS-only feature to work properly. We can
simplify this by consolidating alll the HTTPS redirection logic
here and removing it from Rails.

Closes https://gitlab.com/gitlab-org/gitlab-pages/issues/163