When file not found, possibly redirect to GitLab for acme challenge
In case the file is not found, check the path if it is the Let's
Encrypt acme challenge path. And if so, redirect to gitlab-rails so it
can serve it from the database.

This redirect should only happen when a custom domain is visited, not
the generic GitLab Pages domain. And also HTTP only, no HTTPS.

Closes gitlab-org/gitlab-pages#165