s3cmd: add trailing slash to paths for sync from remote
When backing up the contents of object storage, s3cmd sync will download
the contents of the remote bucket every time even if the files are
available locally. Adding the trailing slash allows s3cmd to reuse
existing local files, which is helpful when using persistent storage for
the backup temp space.

This also allows the --delete-removed option to work (which can be
specified in the .s3cfg); otherwise no files are removed. This in turn
is useful when using a regular registry cleanup, which otherwise would
keep growing the local backup copy unbounded.

See also:
- https://github.com/s3tools/s3cmd/issues/245
- https://s3tools.org/s3cmd-sync