Refactor the code to use `Serving/LookupPath`
This moves the code from the concept
of `Project` to use a concept of `LookupPath`.

This makes the `LookupPath` to define a `Path` on disk that
the data is being served from.

This makes the `ACME` to not have special
handling, but rather try to serve the file
if succeeds, skip GitLab passthrough
for ACME.