Fix make verify ci job
For some reason long awk command makes it crash with a
segmentation fault error.

I tested this by replacing $(allfiles) in
`"Please run ./bin/goimports -w -local $(IMPORT_PATH) -l $(allfiles)"`
with strings of various length, and it fails on longer strings.

I haven't yet figured out the source of the problem, so just
copypasted the approach from workhorse - extracted a separate script
which uses if in shell script and echo to print the suggestion.