Watch a file to configure which domains should use new gitlab source
Instead of passing domain once in an ENV variable we now watcn a file
(specified with `GITLAB_NEW_SOURCE_DOMAINS_FILE`, defaults to
`.new-source-domains`) and read it when it's modified in order to find
which domains should use the new `gitlab` config source.

This way we can control this without having to restart Pages.

Related to https://gitlab.com/gitlab-org/gitlab-pages/issues/266.