Update release helper with upstream desired changes
We want the GitLab Repository to read from a variable

Also add the no-clobber option to some of the copy commands. This is
useful when releasing a patch, where we are specifying the same minor
release folder for our previous release. Otherwise we get a copy error,
about the files not changing.