Support sidekiq-cluster in gitlab-sidekiq image
sidekiq-cluster offers two things:

1. The ability to run and supervise multiple Sidekiq processes,
   potentially listening to different queues.
2. Enhanced queue pre-processing, including the ability to negate a set
   of queues and the experimental query selector.

In this case, we do not want to start multiple Sidekiq processes inside
a single pod - for that, we'd run additional pods. As we only need the
second item, we use the `--dryrun` option to sidekiq-cluster to get the
command line it would execute, and then execute that command directly.