Add configuration flag -enable-domain-source-api
Adds support for `--enable-domain-source-api` boolean default
to `true`. Expects `-gitlab-server` and `-api-secret-key` to be set.
This will be the default behavior for pages. However, it requires
the `/api/v4/internal/pages` endpoint to be enabled on gitlab rails.
A `Ping` will be done to verify the API is enabled.
The daemon will default to get configuration for domains from
disk if the the Ping check fails.