Webservice: convert puma.rb/unicorn.rb to pull from ENV
Move `puma.rb` and `unicorn.rb` from gitlab-webservice to pull their
configuration items from environment variables. This allows us to stop
maintaining multiple copies between codebases (gitlab, Omnibus, CNG, charts).

Combining `ENV` with `||=`, we provide sane defaults based on the Rails
codebase, and the configurations currently generated by the charts.

These changes do not prevent the overiding of these files via mounts,
whether in Docker or Kubernetes. This means it is not a breaking change
to the charts themselves. When the charts make the change to not provide
an overridden configuation, _that_ will be the breaking change.

Part of https://gitlab.com/gitlab-org/charts/gitlab/-/issues/1419