Ensure that we always close GitLab API response body
This change is important because in case of receiving a JSON payload
from GitLab API that is invalid, and is larger than 512, deserializer
will stop reading it after 512 bytes and will not close the reader.

This will result in HTTP connection not being reused.

This commit ensures that we always read reasponse body until EOF and
that we close the reader afterwards.