gitlab-rails: implement rails-db-check
Make use of a tiny Ruby script to query the database and handle
connection to the database. This gives more verbose error handling,
brings us close to the way the actual Rails application will behave,
and still greatly minimizes the time needed to check the running schema
version according to ActiveRecord's handler.

At this time, there is no known-included support for `load_balancing:`
according to
https://docs.gitlab.com/ee/administration/database_load_balancing.html