GraphicsMagick: UBI: populate into rails archive
Populate the contents as part of the gitlab-rails archive output.

Cutting down the content include to binary and copyright (license) to
meet need & licensing requirements. As Rails uses `mini_magick` gem,
and that gem uses `popen3` to call the `gm` binary, development items
(lib, include, pkgconfig) are not needed, and excluded.