ubi8 images: Use LIBDIR instead of hard-coded /usr/lib
The UBI8 builder defines LIBDIR to /usr/lib64, so we should use
that constant to be consistent.

This may be a prerequisite to compile jemalloc with Ruby. Since jemalloc
uses LIBDIR by default, jemalloc libraries are installed in /usr/lib64.
jemalloc does not compile properly with /usr/lib; the configure process
isn't able to determine the length of a `long long`.

Ruby also does not compile properly if jemalloc is in /usr/lib64 and
Ruby is using /usr/lib.