fix(auth): make authentication scope for Pages configurable
This MR makes required authentication permission scope for
Pages configurable.

By default, Pages will use `api` scope to authenticate with
Pages Application registered on GitLab.

With this MR, the scope is configurable and can be set to `read_api`
by providing the `auth-scope` variable in the arguments or in
the `gitlab-pages.conf`

/label ~security

Changelog: added