push_to_redhat.rb pulls from CI_REGISTRY_IMAGE var
The push_to_redhat.rb script has been updated to pull images from the
CI_REGISTRY_IMAGE variable so that tagged builds running on
dev.gitlab.org will find the correct images. In addition, when tagged
builds occur the image tags are constructed from a number of CI
variables  as opposed from the commit reference when the job
runs on gitlab.com (which generally is the branch name).

Signed-off-by: Gerard Hickey <ghickey@gitlab.com>