Use git provided by Gitaly
Historically, it's been quite a hassle to upgrade the git version across
all components of GitLab. This led to diverging git distributions where
we had different versions, patches and compile flags in use in different
parts of the stack.

To fix this problem, we are converting downstream distributions of
Gitaly to use a single source of truth for building git, which is hosted
by the Gitaly project: a simple `make git` will compile the currently
recommended git version. As such, updates of the recommended version
will automatically propagate through the stack. This allows for for a
unified git installation across all of our stack and enables quicker
iteration when e.g. mission-critical patches or security updates are
required or simply when updating the minimum required git version.

Note that this commit also drops all of the custom patches which we
currently have in git. This is because Gitaly has already bumped the
default git version to v2.31.0, which contains all of them.