Upgrade to Go 1.16
Gitaly has the policy of always supporting the latest two stable Go
versions. Following this policy and the release of Go 1.16 in February
2021, this means that it's time to bump the minimum required Go version
for Gitaly to 1.15.

Most projects for which CNG is building images already do support that
version or are in the process of adding tests for that version:

    - gitlab-shell: b49c1d4 (Compile against go 1.14-1.16, 2021-03-13)
    - gitlab-pages: 48ec6de (Set Go 1.16.2 for project, 2021-03-25)
    - gitlab-container-registry: 32b58091 (Bump default Go version to 1.16, 2021-02-22)
    - gitaly: gitaly!3352
    - gitlab-elasticsearch-indexer: gitlab-elasticsearch-indexer!103

Update the Go base image to version 1.16.3.