Build gitlab-pages Via Nix
As this is the simplest golang project we have, I figured it would be a
good first attempt to show how easy it can be to take a project and turn
it into something built by nix.

I think the big scary build is gitaly, but this was _shockingly_ easy,
so I have high hopes.