gitlab-rails: checks/postgresql detect silent NoDatabaseError
`ActiveRecord::ConnectionHandling.postgresql_connection(config)` will
emit `ActiveRecord::NoDatabaseError` when `PG::Error` includes the name
in `production.database` of `database.yml`. This is to indicate a
problem accessing the database shema, but not the database server.

This error is silenced by `ActiveRecord::MigrationContent.current_version`
which results in a `nil` return. This `nil` is not the same as `0`, and
needs to be treated differntly, as it indicated that we could not
determine the version of the schema within the database, because we were
unable to access the configured database name.

Found in https://gitlab.zendesk.com/agent/tickets/212382

References:

- https://github.com/rails/rails/blob/v6.0.3.6/activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb#L48-L54
- https://github.com/rails/rails/blob/v6.0.3.6/activerecord/lib/active_record/migration.rb#L1090-L1093

Changelog: Added