Raise Sidekiq job timeout from 4 to 25 seconds
As described in https://github.com/mperham/sidekiq/wiki/Signals#term,
Sidekiq by default will wait 25 seconds for jobs to terminate after
receiving a TERM signal. The current default of 4 seconds is a bit too
low. Although this value is actually set in Charts, we are setting in
the containers to be consistent.

The change mirrors the Charts update in
https://gitlab.com/gitlab-org/charts/gitlab/-/merge_requests/2249.