Fix Linguist not working on Gitaly nodes
https://gitlab.com/gitlab-org/build/CNG/-/merge_requests/824 changed the
build process in such a way that the default Ruby binstubs (e.g. in
`/usr/bin`) that were installed with `bundle install` are no longer
copied. Instead, binstubs are manually installed in
`/srv/gitlab/bin`. However, Gitaly executes a script that assumes
`git-linguist` is in the `PATH`.

To fix this problem, we append the `/srv/gitlab/bin` to the `PATH`.

Relates to https://gitlab.com/gitlab-org/charts/gitlab/-/issues/3034

Changelog: fixed