Remove gitlab-shell-libproxyproto image
This drops support for the custom gitlab-shell container with
libproxyproto support added in
https://gitlab.com/gitlab-org/build/CNG/-/merge_requests/657.

We no longer need a custom gitlab-shell built with libproxyproto, as
this container is not used and gitlab-sshd is almost in production
(https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/603).

This reverts commit 43b50e73c70cdf5d5b988f4d82992471e370ac7b, reversing
changes made to ee6dc58b053dffde78081141551644486896c143.