fix: fix metrics and logs not including domain resolution time
Currently we do logging and metrics capturing
after we did the domain information lookup.

It allows us to add more information to access logs.

But it also distorts metrics because domain information
lookup takes time.

This logic was originally introduced in
https://gitlab.com/gitlab-org/gitlab-pages/-/merge_requests/157/diffs

It didn't matter back than because we didn't lookup domain via API as we do now.
Now it does matter.

So this commits moves metrics and logging middlewares almost to the top of pipeline.

Changelog: fixed