fix(gitlab-pages): graceful termination
What
---
Move the gitlab-pages process to be PID 1 inside of the container by using `exec` in start up script
and CMD in exec form as well.

Why
---
In https://gitlab.com/gitlab-org/charts/gitlab/-/issues/3249 we found
out that most image we build their main process doesn't run as PID 1.
This can lead to issues where we assume `SIGTERM` is being sent to the
process but it isn't, it's sent to the bash script that starts the
process which doesn't redirect the signal to the gitlab-pages process.

Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>
Reference: https://gitlab.com/gitlab-org/charts/gitlab/-/issues/3249