Make gitlab-go use the Go 1.17.10 tag
Previously for the golang-fips compiler we checked out a branch and a
specific SHA. We had to avoid a shallow clone because the branch could
be updated, leaving out that SHA.

The `go1.17.10-1-openssl-fips` tag was pushed 10 days ago. We can now
use that.

Relates to https://gitlab.com/gitlab-org/charts/gitlab/-/issues/3234

Changelog: changed