Fix owner of public/uploads directory in UBI-8 Sidekiq image
The owner of the `public/uploads` directory on the `gitlab-sidekiq-ee`
UBI-8 image was owned by `root` instead of `git`. The directory was
being created via `mkdir -p public/uploads`, but the owner was never
set properly in the `gitlab-rails` `Dockerfile`. It was set properly
on `gitlab-webservice-ee`, however.

Relates to https://gitlab.com/gitlab-org/charts/gitlab/-/issues/3370

Changelog: fixed