Chore: log ZIP archive corruption error
When a ZIP file cannot be opened for some reason, we previously would
discard the error after attempting to retrieve it from the
cache. While this error is logged in another message, it would be
helpful to have the same message tied with the "archive corrupted" log
entry.

Relates to https://gitlab.com/gitlab-org/gitlab-pages/-/issues/690

Changelog: changed