Fix data race with lookup paths
Since the API response for domains is cached, it's possible that
multiple clients for the same host may cause a data race when sorting
the lookup paths. To ensure this doesn't happen, we sort the lookup
paths once before we cache it.

Relates to https://gitlab.com/gitlab-org/gitlab-pages/-/issues/646

Changelog: fixed