CI: Ensure file permissions are respected
Ensures that file permissions are not modified by the GitLab Runner's
umask setting. This was causing file permissions to change unexpectedly,
including some configuration files being COPY-ed in with 0666 instead of
0644 as specified with the `install` commands.

Related to:
- https://gitlab.com/gitlab-org/gitlab-runner/-/issues/1736
- https://gitlab.com/gitlab-org/charts/gitlab/-/issues/3797

Feature flag is documented here:
https://docs.gitlab.com/runner/configuration/feature-flags.html#available-feature-flags:

>>>
If enabled will remove the usage of umask 0000 call for jobs executed with docker
executor. Instead Runner will try to discover the UID and GID of the user configured
for the image used by the build container and will change the ownership of the working
directory and files by running the chmod command in the predefined container (after
updating sources, restoring cache and downloading artifacts). POSIX utility id must be
installed and operational in the build image for this feature flag. Runner will execute
id with options -u and -g to retrieve the UID and GID.
>>>

Changelog: fixed