Document the why behind the CNG build architecture
The Cloud Native GitLab containers use a complex build
architecture designed for optimizations which may not
be readily apparent. This merge request adds documentation
about why the process works the way it does so that
contributors and consumers have a better perspective of
the build system from end to end.