Fix deprecated bundler flags
Previously `bundle install` was called with the `--clean`,
`--deployment`, and `--without` flags, but these are deprecated in
favor of using `bundle config set --local <flag>`.

This is especially important for the `--deployment` flag since the
`bundler` auto-upgrade mechanism will use the system gem path instead
of `vendor/bundle` if this flag isn't set in the correct way.

Relates to https://gitlab.com/gitlab-org/charts/gitlab/-/issues/4210